+++ multilint/__init__.py
@@ -31,7 +31,7 @@ except ImportError:
 
 __author__ = 'Adam Johnson'
 __email__ = 'me@adamj.eu'
-__version__ = '2.3.0'
+__version__ = '2.4.0'
 
 
 def main():